requirements.txt
Django==4.2
djangorestframework==3.14.0
django-phonenumber-field[phonenumberslite]
celery[redis]==5.2.7
celery_singleton==0.3.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
django-redis==5.2.0
flower==1.2.0
#psycopg2==2.9.6
